dl_test/requirements.txt
# 🐕 Dog Similarity Search with Keypoint Analysis
# Requirements file generated from actual environment

# 핵심 딥러닝 패키지
torch==2.7.1+cpu
torchvision==0.22.1+cpu
torchaudio==2.7.1+cpu
numpy==1.24.3

# 컴퓨터 비전 및 키포인트 검출
mmcv==2.1.0
mmpose==1.3.2
mmdet==3.3.0
mmengine==0.10.7

# 웹 프레임워크
fastapi==0.115.14
uvicorn==0.34.3
python-multipart==0.0.20

# 이미지 처리
pillow==11.2.1
opencv-python==4.11.0.86

# 데이터 처리 및 유틸리티
scikit-learn==1.7.0
matplotlib==3.10.3
tqdm==4.67.1
transformers==4.53.0
timm==1.0.16

# COCO 데이터셋 도구
xtcocotools==1.14.3
pycocotools==2.0.10

# 추가 의존성 패키지들
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
chumpy==0.70
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
cython==3.1.2
exceptiongroup==1.3.0
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
h11==0.16.0
huggingface-hub==0.33.1
idna==3.10
jinja2==3.1.4
joblib==1.5.1
json-tricks==3.17.3
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
munkres==1.1.4
networkx==3.3
packaging==25.0
platformdirs==4.3.8
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
setuptools==63.2.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
typing-extensions==4.12.2
typing-inspection==0.4.1
urllib3==2.5.0
wheel==0.45.1
yapf==0.43.0
